refactor(domain): move Dictionary domain to a package
