+++ packages/@proofdict/domain/src/Dictionary/DictionaryAllow.ts
