+++ packages/@proofdict/domain/src/Dictionary/DictionaryAllowFactory.ts
