+++ packages/@proofdict/domain/src/Dictionary/DictionaryAllows.ts
