+++ packages/@proofdict/domain/src/Dictionary/DictionaryDescription.ts
