+++ packages/@proofdict/domain/src/Dictionary/DictionaryPattern.ts
