+++ packages/@proofdict/domain/src/Dictionary/DictionaryPatterns.ts
