+++ packages/@proofdict/domain/src/Dictionary/DictionarySlugCreator.ts
