+++ packages/@proofdict/domain/src/Dictionary/DictionarySpec.ts
