+++ packages/@proofdict/domain/src/Dictionary/DictionarySpecs.ts
