+++ packages/@proofdict/domain/src/Dictionary/DictionaryTags.ts
