+++ packages/@proofdict/domain/src/index.ts
@@ -1,0 +1,14 @@
+export * from "./Dictionary/Dictionary";
+export * from "./Dictionary/DictionaryFactory";
+export * from "./Dictionary/DictionaryDescription";
+export * from "./Dictionary/DictionaryExpected";
+export * from "./Dictionary/DictionaryPattern";
+export * from "./Dictionary/DictionaryPatterns";
+export * from "./Dictionary/DictionaryAllow";
+export * from "./Dictionary/DictionaryAllows";
+export * from "./Dictionary/DictionaryAllowFactory";
+export * from "./Dictionary/DictionarySpec";
+export * from "./Dictionary/DictionarySpecs";
+export * from "./Dictionary/DictionarySlugCreator";
+export * from "./Dictionary/DictionaryTag";
+export * from "./Dictionary/DictionaryTags";
