+++ packages/@proofdict/domain/test/mocha.opts
@@ -1,0 +1,1 @@
+--require ts-node-test-register
