+++ packages/@proofdict/editor/src/component/ui-kit/Grid/GridCell.tsx
@@ -1,4 +1,5 @@
 import * as React from "react";
+
 const classNames = require("classnames");
 require("./GridCell.css");
 export interface GridCellProps {
