+++ packages/@proofdict/editor/src/infra/formatter/PrhFormatter.ts
@@ -1,6 +1,6 @@
 // MIT © 2017 azu
 import * as yaml from "js-yaml";
-import { DictionaryJSON } from "../../domain/Dictionary/Dictionary";
+import { DictionaryJSON } from "@proofdict/domain";
 
 export const modifyForPrh = (json: DictionaryJSON): Object => {
     const description = json.description
