+++ packages/@proofdict/editor/src/infra/prh/__tests__/Prh-test.ts
@@ -1,10 +1,7 @@
 // MIT © 2017 azu
 import { getMatchExpectedWords, getUniqueTokens, testPattern } from "../Prh";
-import { createDictionary } from "../../../domain/Dictionary/DictionaryFactory";
-import { DictionaryPattern } from "../../../domain/Dictionary/DictionaryPattern";
-import { DictionaryExpected } from "../../../domain/Dictionary/DictionaryExpected";
+import { createDictionary, DictionaryExpected, DictionaryPattern, DictionarySpec } from "@proofdict/domain";
 import * as assert from "assert";
-import { DictionarySpec } from "../../../domain/Dictionary/DictionarySpec";
 
 describe("Prh", () => {
     describe("#testPattern", () => {
