+++ packages/@proofdict/editor/src/use-case/dictionary/AddNewSpecToDictionaryUseCase.ts
@@ -1,8 +1,7 @@
 // MIT © 2017 azu
 import { UseCase } from "almin";
 import { dictionaryRepository, DictionaryRepository } from "../../infra/repository/DictionaryRepository";
-import { DictionaryIdentifier } from "../../domain/Dictionary/Dictionary";
-import { DictionarySpec } from "../../domain/Dictionary/DictionarySpec";
+import { DictionaryIdentifier, DictionarySpec } from "@proofdict/domain";
 
 export const createAddNewSpecToDictionaryUseCase = () => {
     return new AddNewSpecToDictionaryUseCase({
