+++ packages/@proofdict/editor/src/use-case/dictionary/ImportDictionaryFromJSONUseCase.ts
@@ -1,6 +1,6 @@
 // MIT © 2017 azu
 import { UseCase } from "almin";
-import { DictionaryJSON, DictionarySerializer } from "../../domain/Dictionary/Dictionary";
+import { DictionaryJSON, DictionarySerializer } from "@proofdict/domain";
 import { dictionaryRepository, DictionaryRepository } from "../../infra/repository/DictionaryRepository";
 import { createUpdateDictionarySpecStatusUseCase } from "./UpdateDictionarySpecStatusUseCase";
