+++ packages/@proofdict/editor/src/use-case/dictionary/UpdateDictionaryDescriptionUseCase.ts
@@ -1,8 +1,7 @@
 // MIT © 2017 azu
 import { UseCase } from "almin";
 import { dictionaryRepository, DictionaryRepository } from "../../infra/repository/DictionaryRepository";
-import { DictionaryIdentifier } from "../../domain/Dictionary/Dictionary";
-import { DictionaryDescription } from "../../domain/Dictionary/DictionaryDescription";
+import { DictionaryDescription, DictionaryIdentifier } from "@proofdict/domain";
 
 export const createUpdateDictionaryDescriptionUseCase = () => {
     return new UpdateDictionaryDescriptionUseCase({
