+++ packages/@proofdict/editor/src/use-case/dictionary/UpdateDictionarySpecStatusUseCase.ts
@@ -1,9 +1,8 @@
 // MIT © 2017 azu
 import { UseCase } from "almin";
 import { dictionaryRepository, DictionaryRepository } from "../../infra/repository/DictionaryRepository";
-import { DictionaryIdentifier } from "../../domain/Dictionary/Dictionary";
+import { DictionaryIdentifier, DictionarySpecs } from "@proofdict/domain";
 import { testPattern } from "../../infra/prh/Prh";
-import { DictionarySpecs } from "../../domain/Dictionary/DictionarySpecs";
 
 export const createUpdateDictionarySpecStatusUseCase = () => {
     return new UpdateDictionarySpecStatusUseCase({
